node-statusmon is now in npm as 'statusmon'.
Required some changes to ensure the probes could be loaded, we
now have a static list in the config file, as I can't figure out
how to retain the current behaviour with npm.

Eventually this will need some updates anyway so that it's easy
to support user modules, so we can revisit it then.